:recycle: rename school's application elements
