+++ api/src/school/application/assessment-controller.js
@@ -1,6 +1,6 @@
-import * as challengeSerializer from '../../../../lib/infrastructure/serializers/jsonapi/challenge-serializer.js';
-import { usecases } from '../../shared/usecases/index.js';
-import * as assessmentSerializer from '../../../../src/school/infrastructure/serializers/assessment.js';
+import * as challengeSerializer from '../../../lib/infrastructure/serializers/jsonapi/challenge-serializer.js';
+import { usecases } from '../shared/usecases/index.js';
+import * as assessmentSerializer from '../../../src/school/infrastructure/serializers/assessment.js';
 
 const getNextChallengeForPix1d = async function (request, h, dependencies = { challengeSerializer }) {
   const assessmentId = request.params.id;
