+++ api/src/school/application/organization-learner-route.js
@@ -1,7 +1,7 @@
 import Joi from 'joi';
-import { organizationLearnerController } from './controller.js';
-import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
-import { identifiersType } from '../../../../lib/domain/types/identifiers-type.js';
+import { organizationLearnerController } from './organization-learner-controller.js';
+import { securityPreHandlers } from '../../../lib/application/security-pre-handlers.js';
+import { identifiersType } from '../../../lib/domain/types/identifiers-type.js';
 
 const register = async function (server) {
   server.route([
