+++ api/src/school/application/school-controller.js
@@ -1,5 +1,5 @@
-import { usecases } from '../../shared/usecases/index.js';
-import * as schoolSerializer from '../../infrastructure/serializers/school-serializer.js';
+import { usecases } from '../shared/usecases/index.js';
+import * as schoolSerializer from '../infrastructure/serializers/school-serializer.js';
 
 const getSchool = async function (request) {
   const { code } = request.params;
