+++ api/src/school/application/school-route.js
@@ -1,7 +1,7 @@
 import Joi from 'joi';
-import { schoolController } from './controller.js';
-import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
-import { identifiersType } from '../../../../lib/domain/types/identifiers-type.js';
+import { schoolController } from './school-controller.js';
+import { securityPreHandlers } from '../../../lib/application/security-pre-handlers.js';
+import { identifiersType } from '../../../lib/domain/types/identifiers-type.js';
 
 const register = async function (server) {
   server.route([
