+++ api/src/school/routes.js
@@ -1,6 +1,6 @@
-import * as assessment from './application/assessments/route.js';
-import * as organizationLearner from './application/organization-learner/index.js';
-import * as school from './application/school/index.js';
+import * as assessment from './application/assessment-route.js';
+import * as organizationLearner from './application/organization-learner-route.js';
+import * as school from './application/school-route.js';
 const schoolRoutes = [assessment, school, organizationLearner];
 
 export { schoolRoutes };
