+++ api/tests/school/integration/application/assessment-controller_test.js
@@ -1,7 +1,7 @@
-import { expect, hFake, sinon } from '../../../../test-helper.js';
-import { assessmentController } from '../../../../../src/school/application/assessments/controller.js';
-import { usecases } from '../../../../../src/school/shared/usecases/index.js';
-import { Assessment } from '../../../../../src/school/domain/models/Assessment.js';
+import { expect, hFake, sinon } from '../../../test-helper.js';
+import { assessmentController } from '../../../../src/school/application/assessment-controller.js';
+import { usecases } from '../../../../src/school/shared/usecases/index.js';
+import { Assessment } from '../../../../src/school/domain/models/Assessment.js';
 
 describe('Integration | Controller | assessment-controller', function () {
   describe('#createForPix1d', function () {
