+++ api/tests/school/unit/application/assessment-controller_test.js
@@ -1,6 +1,6 @@
-import { expect, hFake, sinon } from '../../../../test-helper.js';
-import { assessmentController } from '../../../../../src/school/application/assessments/controller.js';
-import { usecases } from '../../../../../src/school/shared/usecases/index.js';
+import { expect, hFake, sinon } from '../../../test-helper.js';
+import { assessmentController } from '../../../../src/school/application/assessment-controller.js';
+import { usecases } from '../../../../src/school/shared/usecases/index.js';
 
 describe('Unit | Controller | assessment-controller', function () {
   describe('#getNextChallengeForPix1d', function () {
