+++ api/tests/school/unit/application/organization-learner-controller_test.js
@@ -1,7 +1,7 @@
-import { expect, hFake, sinon } from '../../../../test-helper.js';
-import { usecases } from '../../../../../src/school/shared/usecases/index.js';
-import { organizationLearnerController } from '../../../../../src/school/application/organization-learner/controller.js';
-import { OrganizationLearner } from '../../../../../src/school/domain/models/OrganizationLearner.js';
+import { expect, hFake, sinon } from '../../../test-helper.js';
+import { usecases } from '../../../../src/school/shared/usecases/index.js';
+import { organizationLearnerController } from '../../../../src/school/application/organization-learner-controller.js';
+import { OrganizationLearner } from '../../../../src/school/domain/models/OrganizationLearner.js';
 
 describe('Unit | Controller | organization-learner-controller', function () {
   describe('#getById', function () {
