+++ api/tests/school/unit/application/school-controller_test.js
@@ -1,8 +1,8 @@
-import { expect, hFake, sinon } from '../../../../test-helper.js';
-import { usecases } from '../../../../../src/school/shared/usecases/index.js';
-import { schoolController } from '../../../../../src/school/application/school/controller.js';
-import { School } from '../../../../../src/school/domain/models/School.js';
-import { OrganizationLearner } from '../../../../../src/school/domain/models/OrganizationLearner.js';
+import { expect, hFake, sinon } from '../../../test-helper.js';
+import { usecases } from '../../../../src/school/shared/usecases/index.js';
+import { schoolController } from '../../../../src/school/application/school-controller.js';
+import { School } from '../../../../src/school/domain/models/School.js';
+import { OrganizationLearner } from '../../../../src/school/domain/models/OrganizationLearner.js';
 
 describe('Unit | Controller | school-controller', function () {
   describe('#getSchool', function () {
